fixed bit tlg ids

+++ persist/user_locales.sql
@@ -1,6 +1,6 @@
 CREATE table idle_rpg_bot.user_locales
 (
-  telegram_id integer,
+  telegram_id bigint,
   locale      text
 );
 alter table  idle_rpg_bot.user_locales owner to idle_rpg_bot;
